requirements.txt
pyhamcrestaenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
allure-pytest==2.14.0
allure-python-commons==2.14.0
Appium-Python-Client==5.1.1
attrs==25.3.0
behave==1.2.6
certifi==2025.1.31
charset-normalizer==3.4.1
dill==0.4.0
execnet==2.1.1
Faker==37.1.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
playwright==1.54.0
pluggy==1.5.0
propcache==0.3.2
pyee==13.0.0
PyHamcrest==2.1.0
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
pytest-reportportal==5.5.2
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dotenv==1.1.0
python-slugify==8.0.4
reportportal-client==5.6.5
requests==2.32.3
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
